Add LOCAL vs PREINSTALLED selection for `add_to_settings_path()`
The method `cc.core.paths.add_to_settings_path()` allowed a new search folder to be added to the very beginning or the very end; however, it is more frequently useful to add it at the end of the LOCAL settings folders but before the PREINSTALLED ones.

This breaks up the `settings_path()` list into two separate components,
`local_settings_path()` and `preinstalled_settings_path()`, each of which may be
modified.